requirements.txt
streamlit==1.33.0
Pillow==10.2.0
openai==1.14.3
tflite-runtime
selenium==4.18.1
webdriver-manager==4.0.1
numpy==1.26.4
huggingface_hub==0.22.2
python-dotenv==1.0.1
